Cambio en protocolo para el send/receive de posicion mal mapeada. Se agrega el seteo del modo en la partida/juego.

+++ common_libs/include/common_estado_juego.h
@@ -11,8 +11,8 @@ private:
     uint8_t game_over;
     uint8_t id;
     uint8_t id_character; //revisar el valor
-    uint8_t x;
-    uint8_t y;
+    uint32_t x;
+    uint32_t y;
     uint8_t ABM;
     uint8_t accion;
     uint8_t vida;
